moved addPositions to services file, position isn't added unless applicationsOpen is true
