+++ functions/src/Controller/Elections.ts
@@ -1,8 +1,7 @@
 import { electionsConverter } from '../Models/Elections';
 import * as functions from 'firebase-functions';
 import { db } from '../index';
-import { openElections, addPosition } from '../Services/Elections';
-import { Position } from '../Models/Position';
+import { openElections } from '../Services/Elections';
 
 export function getElectionsCollection() {
 	return db.collection('elections').withConverter(electionsConverter);
@@ -10,12 +9,5 @@ export function getElectionsCollection() {
 
 export const openElectionsController = functions.https.onRequest((request, response) => {
 	openElections();
-	response.status(200).send('Good Job');
-});
-
-export const addPositionController = functions.https.onRequest((request, response) => {
-	const position: Position = new Position(request.body);
-
-	addPosition(position);
 	response.status(200).send('Good Job');
 });
